Split issues.tsx into two files.
